requirements.txt
openai
python-dotenv
streamlit
pypdf
ebooklib
bs4